Java program

+++ javaPractice/com/Banking/PeerToPeerTransaction.java
@@ -2,7 +2,7 @@ package com.Banking;
 class  PeerToPeerTransaction extends Transaction{
     private String recipient;
 
-    public PeerToPeerTransaction(int transactionId, long amount, String recipient)
+    public PeerToPeerTransaction(int transactionId, double amount, String recipient)
     {
         super(transactionId,amount);
         this.recipient = recipient;
@@ -10,4 +10,8 @@ class  PeerToPeerTransaction extends Transaction{
      String getRecipient(){
         return recipient;
     }
+    //   public PeerToPeerTransaction()
+    // {
+
+    // }
 }
